requirements.txt
openai
python-dotenv
transformers
torch
accelerate
google-genai
google-api-core
